Move utility functions to runner.dart
